Move out logic from AsyncHandlerMethods to NingWSCache.
